+++ utils/testutils/test_helpers.go
@@ -1,4 +1,4 @@
-package utils
+package testutils
 
 import (
 	"io/ioutil"
